api/requirements.txt
# Core FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning and Data Processing
faiss-cpu==1.11.0.post1
numpy<2.0
Pillow==10.1.0
opencv-python-headless==4.9.0.80
transformers==4.38.2
torch==2.2.1
timm==0.9.16

# Data validation and serialization
pydantic==2.4.2

# Built-in Python modules (no installation required)
# json - built-in
# os - built-in
# math - built-in
# hashlib - built-in
